some cleanups, added range checks. (bst_log_adjustment_setup): turn
Fri Oct 24 22:29:02 2003  Tim Janik  <timj@gtk.org>

        * beast-gtk/bstlogadjustment.[hc]: some cleanups, added range checks.
        (bst_log_adjustment_setup): turn n_steps into a double.
        emit ::changed and ::value-changed after updating ranges.

svn path=/trunk/; revision=1433

+++ beast-gtk/gxk/gxklogadjustment.h
@@ -66,7 +66,7 @@ GtkAdjustment*	bst_log_adjustment_from_adj	(GtkAdjustment		*client);
 void		bst_log_adjustment_setup	(BstLogAdjustment	*ladj,
 						 gdouble		 center,
 						 gdouble		 base,
-						 guint			 n_steps);
+						 gdouble		 n_steps);
 
 #ifdef __cplusplus
 }
